+++ packages/sidecar/CHANGELOG.md
@@ -1,8 +1,8 @@
-# @kalguard/sidecar
+# kalguard-sidecar
 
 ## 0.1.1
 
 ### Patch Changes
 
 - Updated dependencies []:
-  - @kalguard/core@0.2.0
+  - kalguard-core@0.2.0